Move some of the subgraph extraction functions into own module
